fix: move strokes updates to happen on moue rather than via useEffect
React docs recommend doing update logic on the event,
not using a state variable inbetween.
This was how useLiveUpdates worked previously,
so we switched it to returning a useCallback for updateStrokes rather than a state setter.
This was to prevent an infinite loop from setLatex changing,
and calling itself on change which then caused it to change again.